Fix loading of a kernel that was linked with ld 2.20 on x86_64.
The MAXPAGESIZE changed between binutils 2.17 and 2.20 from 0x100000
to 0x200000 for x86_64. But we have 0x100000 coded into our ldscript
for the kernel.

The real solution would be to specify "CONSTANT (MAXPAGESIZE)" in
the ldscript but unfortunately binutils 2.17 (at least our version;
later snapshots like 2.17.50 or 2.17.90 do) doesn't have support for
this keyword.

We will do that too once we get rid of binutils 2.17. Until then,
set max-page-size to 0x100000 as a workaround when linking on
x86_64.