Fix buildkernel without -j.
One of the recent commits by sephe caused genassym.c to depend on
bus_if.h and device_if.h (via machintr.h via bus.h).

However, without -j, these headers would not be generated (from
the .m files) before genassym.c was compiled, causing the kernel
build to break.

As a fix, make genassym.c depend on these headers in the Makefile.

The interesting bit is that this is one of the cases where it
breaks without -j but _does_ work with -j 1 (read: one) due to
the different handling of no -j vs. -j 1 in make(1).