xlint: Revert part of 2e2a1cf79437aa99c731a36585fe98887a5c0bf0 again.
The reason why we had to lower WARNS to 0 for xlint/lint1 was that
gcc 4.4.2 was unable to fit the value for LDBL_MAX from <float.h>
into a long double on i386 due to some misconfiguration of the
compiler.

Now that this is fixed with gcc 4.4.5, we can remove it again.