sh(1): Describe subshell environment, command substitution better
POSIX does not require the shell to fork for a subshell environment, and
we use that possibility in various ways (command substitutions with a
single command and most subshells that are the final command of a shell
process).  Therefore do not tie subshells to forking in the man page.

Also, fix some sentences with excessive semicolons.

Obtained-from:  FreeBSD