kernel: Implement ELF auxinfo entry type AT_EXECPATH
Refer to issue 2005.
The new RTLD can optionally use AT_EXECPATH for some of its new
functionality such as libmap.conf and dynamic token expansion.
In the absence of this kernel support, RTLD has a fallback
mechanism that works most of the time, but using it takes a
performance hit.

This patch was tested for some time by various people prior to
the 2.10 branch being created.