sort: modifications needed for DragonFly
Minor modications are needed to get NetBSD's sort to build on DragonFly
without error.  The emalloc and erealloc functions are actually now
supported on DragonFly within libutil, but they can't be used because
sort is one of the bootstrap tools.  During the bootstrap, libutil is
not available, so the original malloc and realloc function have been
used along with err() to simulate these error-checking functions without
requiring libutil.