grep: Delete redundant headers
It was believed that all the headers deleted in this commit were generated
by the vendor makefile.  The commit message
<http://leaf.dragonflybsd.org/mailarchive/commits/2011-04/msg00216.html>
highlighted that they were actually part of the source tarball.  The
various makefiles have been altered to look for those instead, and
duplicate headers such as config.h, configmake.h, and arg-nonnull.h were
consolidated as well.