libutil: Create a <util.h> symlink.
As stuff from NetBSD goes into our libutil, pkgsrc starts detecting that
and assumes we have a <util.h> header, too. But ours is named <libutil.h>
due to FreeBSD heritage.

In an attempt to make life easier and not have to patch packages for this
every time, symlink <util.h> to <libutil.h> for now.

However, it could well be that other packages test for the existence of
<util.h> and (upon finding it) assume we have some NetBSD function in
libutil which we don't have. But we'll see.

As-proposed-by: y0netan1