kernel/ddb: Fix examine's z format (for signed hex).
This was previously attempted to be printed with a special %z conversion
specifier of kprintf(). But our kprintf() doesn't support this specifier.
Either support for it got lost along the way or we never had it, I didn't
check.

Since we want to get rid of non-standard conversion specifiers (because
they require compiler modification for silencing format warnings), use a
db_format_hex() function that does the same as %z was supposed to do.

Taken-from: NetBSD