cvs: fix recently introduced, repeatable segfault
In specific involving local cvs repositories, the newly refactored cvs
would segfault.  It was traces back to libc's version of delim.  The
previous version of cvs provided its own functions like getline and
delim, but the recent change choose to use libc instead.

This "fix" basically goes back to the provided version of delim.  There
might be something wrong with the libc version.  It would work for a
while, but eventually it segfaulted on a memcpy command.

reported-by: swildner