gcc 44: Change shared library search order (embedded rpath)
GCC 4.4.x has always created shared libraries with embedded rpath of
/usr/lib/gcc44:/usr/lib.  There are only two shared libraries in
/usr/lib/gcc44 which are stdc++ and obj-c.  There are 73 shared libraries
at /usr/lib, including the c library.  Reversing the rpath order will
cause many more library searchs to be successful on the first hit.

When searching for a library created by pkgsrc, there is no difference.
In that case, /usr/pkg/lib becomes the third directory in the search
path and relative order of the first two is irrelevant as both fail.