btx: tweak makefile for gold linker
gold linker ignores the -i flag, which is considered a synonym for -r
flag (--relocatable) by ld linker.  This causes the btx build of crt0
to fail under gold.  The easy solution changes the -i flag to -r since
both linkers understand it.

This is the only instance of "-i" in the system.  All the other products
that require the relocatable switch use the "-r" option, so now there
are no "-i" flags left.