gold: Local modification to add --relocatable synonym
Until a change to the binutils master branch two days ago, the gold linker
simply ignored the -i switch, which is considered to be a synonym of
the --relocatable (-r) switch by the ld linker.

This mainly affects makefiles that use gold directly rather than having it
invoked by gcc, and there was a case like this in the DragonFly base.
While the base was easily modified to use the -r switch, it was deemed a
good idea to apply this modification to the gold of binutils 2.21 in case
any pkgsrc software try to use the linker directly.