Kernel Elf: Expand ABI.note-tag search to cover entire file
The FreeBSD code that provides the new ABI.note-tag search limits itself
to the first page PT_NOTE header found and there could be more than one.
Moreover, that header is ignored if it doesn't fall within the first page.

The problem is that 99% of the binaries produced right now have their
note-tag outside of the first page, and perhaps the culprit is the new
binutils 2.21.  Due to this, we are required to search the entire
binary to look for these tags, and the previous code of DragonFly did
this so part of it was brought back here.