kernel ELF: Implement AT_OSRELDATE aux vector
We have several unimplemented non-standard enumerations of the ELF aux
vectors brought in with a recent sync with FreeBSD.  In general they won't
be implemented because of philosophical reasons or lack of necessity.  In
the case of AT_OSRELDATE, it makes sense to provide this simple update
as the release date comes directly from the ELF binary and it's already
available thanks to recent comments regarding osreldate.