i386/lapic: Allow vmware with single CPU to boot SMP kernel
It looks like that if only one CPU is configured in vmware, then
LAPIC will not work at all, so LAPIC will not be enabled if this
kind of configuration is detected.

Currently we detect vmware using cpu_feature2's CPUID2_VMM bit.

Thank swildner@ for providing necessary information on this.