kernel/lapic: Remove the special handling for VMware w/ 1 CPU again.
I've done some further testing here and it looks like I was on the
wrong path when I reported it to sephe.

The problem seems to be a local one. It fails occasionally (SMP kernel
or not) when the ISO is on my file server, but always succeeds when
the ISO is on the disk of the box running VMware.