vga(4): Remove a bogus check.
These are the remains of a larger chunk of code that I removed in
4e193d20. It was used to determine whether a MGA, CGA, EGA or VGA
was present in the box. I left the check in at the time because I
thought, why not?

However, ASpeed's VGA BIOS (at least as found on some Tyan S8230*
series motherboards) seems to have some flaws and ends up returning
0xf instead of 0x9. This caused vga(4) initialization to fail early
in the boot and no kernel messages would go to the screen.

Since the only thing that interests us is whether a VGA is present
or not, remove this check. There are other tests later on that will
detect what we want (such as checking the CRTC's address).

Reported-by: Francois Tigeot <ftigeot@wolfpond.org>