BSD gprof/4: Remove old gprof versions from code base
Gprof4 was inherited from FreeBSD, but removed from there in March 2002.
It was supposed to handle 8-byte counters which the standard gprof didn't
do at the time, but was later upgraded.  Gprof4 probably should not have
made it to DragonFly 1.0.  :)

BSD gprof has been replaced by GNU gprof, so let's retire it formally.