ioapic: Don't do the default IRQ configure, if it is mapped to different GSI
The problem went on unnoticed for such a long time, mainly because most of
the mobos only map IRQ0 to GSI2, and GSI0 is taken as being hooked to the
external interrupt source like i8259.

However, the bug reporter's mobo has an extra IRQ source override entry in
ACPI MADT, which maps IRQ9 to GSI21.  The problem reported is actually
caused by:
GSI is configured in the ascending order, so GSI9 is configured before GSI21.
Since there is no explicit IRQ to GSI mapping for GSI9, by using the default
1:1 mapping IRQ9 will be configured.  When GSI21 is configured, due to the
explicit IRQ to GSI mapping, IRQ9 will be configured again, booo

Reported-by: Siju George <sgeorge.ml@gmail.com>