gdb(1): Fix the dreaded "xml-builtin.c already exists" quickworld error.
The feature-to-c.sh script has a check that outputs this error if the
file exists but would have to be regenerated (as can be the case when
quickworlding).

Because fixing it in the script would take more lines than fixing it
in the Makefile, I do the latter (the effect is the same).