ethernet: Add ether_demux(), which will send the packet to the correct netisr
It is intended to be used by the function which does not know whether
the current thread is the correct netisr for the ethernet packet or not.

Currently only ng_ether uses this function, which incorrectly called
ether_demux_oncpu().  The other two callers of ether_demux_oncpu() are
already on the correct netisr.

This paves the way to avoid requeuing the packet to the current netisr
in the ether_demux_oncpu().