Revert "libcrypt - install crypt.h header"
This reverts commit b4ed82ece2b69f4a6711d35c5e42938dfc1d804c.

BSDs have libcrypt and the prototypes for its functions are in
<unistd.h>. The reason we had crypt.h installed for a while was
to make KDE link against libcrypt, due to a wrong check in KDE.

Unfortunately, at least one other package (chat/dircproxy)
assumed that if <crypt.h> exists, it would also find prototypes
for crypt() and friends there, which is not the case. So it
would crash on x86_64 due to defaulting to int as crypt()'s
return type (which is a pointer).

The check in KDE has been fixed since and it properly checks
for the presence of libcrypt:

https://bugs.kde.org/show_bug.cgi?id=247627

Hence this revert.

In-discussion-with: alexh