Binutils 2.20: Effectively remove from world
There is no longer a reason to maintain multiple versions of binutils
in the base system.  While contrib/binutils-2.20 directory isn't being
removed quite yet, this commit effectively removed binutils 2.20 from
DragonFly.

Sometime in the future, binutils may be removed from the objformat
handler.  The value of the BINUTILSVERS variable no longer has any
effect, and the only version of binutils on the system is 2.21.