libmpfr: Upgrade to version 3.1.0 from version 2.4.2
Like libgmp, libmpfr has the ability to tune itself to the cpu used, e.g.
core2, k8, etc.  Currently we are not taking advantage of this situation,
but in the future the CPU_TYPE variable should be examined for tuning
purposes.

libmpfr is a requirement to build gcc-4.4 and later versions.