acpi_cpu: do not apply P_LVLx_LAT rules to latencies returned by _CST
ACPI specification states that if P_LVL2_LAT >100, then a system
doesn't support C2; if P_LVL3_LAT >1000, then C3 is not supported.

But there are no such rules for Cx state data returned by _CST. If a
state is not supported it should not be included into the return
package. In other words, any latency value returned by _CST is valid,
it's up to the OS and/or user to decide whether to use it.

Submitted-by:  Johannes Hofmann <Johannes.Hofmann@gmx.de>
Taken-from:    FreeBSD (rev 212549)
Dragonfly-bug: <http://bugs.dragonflybsd.org/issue2157>