Update the pciconf(8) database.
Dec 21 snapshot from http://pciids.sourceforge.net/

This commit switches us to using the sourceforge list only. It's large,
actively maintained, and comes in the right format for pciconf(8) so we
need no scripts to prepare it.

Previously it was a combination of this database with Chris Hart's (seems
to be no longer actively maintained) and Jim Boemler's.