kgdb: Fix double-free bug
It was a mistake to reuse target sections for each inferior.  Leaving the
default section for each program space eliminates the redundant xfrees.
The exec.c function "resize_table_sections" code was painful to look at with
unused variables and unnecessary memory reallocations, so it was altered
at the same time.