gold linker v2.22: Fix constructor handling
Starting with gold 2.22, gold looks for contructors in DT_INIT_ARRAY by
default, but gcc defines them in a .ctors section.  Apparently the
presence of crtbegin and crtend with their .init sections are confusing
gold, but not gnu ld.

To combat this, a new option "--no-ctors-in-init-array" was added to
emulate gnu ld behavior of looking for constructors in the .ctors
section provided by gcc.  This commit makes this new option the default
behavior on DragonFly's gold.

Without this modification, a gold-built world is unbootable!