rtld: Clang fixes
1) Fix clang warnings
2) Fix r_debug_state function

The r_debug_state function is a hook required for gdb.  Clang will inline
the function, but because it is empty, it won't generate any calls to it.
This results in gdb backtraces being corrupted.  Fixed with the following
hack obtained from FreeBSD SVN 225366.

Fix it by marking the function __noinline, and inserting an empty asm
statement that pretends to clobber memory.  This forces the compiler to
emit calls to the r_debug_state() function throughout rtld.c