kern: Update traps, sigbus->sigsegv, cleanup and fixes
The primary purpose of this changeset is to change the signal from
bus error to segfault when an attempt to access protected memory is
made.  In the process of doing this, it was noticed that there were
differences between i386 and x86_64 as well as differences between
the actual and virtual kernels.

Several of these differences were addressed.  Extra whitespace was
removed and some syncing between the divergent FreeBSD version was
made.  Some "#if 0" sections were removed.  The virtual kernels were
tested along with the real kernels.  A conservative approach was
taken as there seems to more cruft that could come out.